+++ runtime/configs/config.go
@@ -3,8 +3,9 @@ package configs
 import "github.com/pokt-network/pocket/runtime/defaults"
 
 type Config struct {
-	RootDirectory string `json:"root_directory"`
-	PrivateKey    string `json:"private_key"` // INVESTIGATE(#150): better architecture for key management (keybase, keyfiles, etc.)
+	RootDirectory   string `json:"root_directory"`
+	PrivateKey      string `json:"private_key"` // INVESTIGATE(#150): better architecture for key management (keybase, keyfiles, etc.)
+	ClientDebugMode bool   `json:"client_debug_mode"`
 
 	Consensus   *ConsensusConfig   `json:"consensus"`
 	Utility     *UtilityConfig     `json:"utility"`