+++ shared/modules/bus_module.go
@@ -6,6 +6,8 @@ import (
 	"github.com/pokt-network/pocket/shared/messaging"
 )
 
+const BusModuleName = "bus"
+
 // DISCUSS if this channel should be of pointers to PocketEvents or not. Pointers
 // would avoid doing object copying, but might also be less thread safe if another goroutine changes
 // it, which could potentially be a feature rather than a bug.
@@ -16,6 +18,8 @@ type Bus interface {
 	PublishEventToBus(e *messaging.PocketEnvelope)
 	GetBusEvent() *messaging.PocketEnvelope
 	GetEventBus() EventsChannel
+
+	RegisterModule(module Module) error
 
 	// Pocket modules
 	GetPersistenceModule() PersistenceModule