+++ p2p/CHANGELOG.md
@@ -7,9 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.15] - 2023-01-09
+## [0.0.0.18] - 2023-01-10
 
 - Add a lock to the mempool to avoid parallel messages which has caused the node to crash in the past
+
+## [0.0.0.17] - 2023-01-10
+
+- Updated module constructor to accept a `bus` and not a `runtimeMgr` anymore
+- Registering module with the `bus` via `RegisterModule` method
+- Updated tests and mocks accordingly
+- Sorting `validatorIds` in `testRainTreeCalls`
+
+## [0.0.0.16] - 2023-01-09
+
+- Added missing `Close()` call to `persistenceReadContext`
+
+## [0.0.0.15] - 2023-01-03
+
+- Refactored `AddrBookProvider` to support multiple implementations
+- Added `CurrentHeightProvider`
+- Dependency injection of the aforementioned provider into the module creation (used by the debug-client)
+- Updated implementation to use the providers
+- Updated tests and mocks
 
 ## [0.0.0.14] - 2023-01-03